Better error messages for golang wrapper dependency issues
These error messages contain shell commands which should be run in order to
bring the golang wrappers into line with the current solidity contracts and the
current geth version. Those instructions were easy to miss in the error output,
which was blocking people. This changes the output highlight the instructions.